pyannote-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pyannote.audio>=3.1.0
torch>=2.1.0
torchaudio>=2.1.0
python-multipart>=0.0.6
